refactor: extract shared getSeroSettings and readSettings helpers
