+++ apps/desktop/electron/shared/settings/model-fallback-chain.ts
@@ -5,6 +5,8 @@
  * sensible default chain if the setting is missing so users can override it.
  */
 
+import { getSeroSettings } from './settings-helpers';
+
 const DEFAULT_FALLBACK_CHAIN = [
   'gpt-5.4',
   'gpt-4.1-mini',
@@ -30,14 +32,6 @@ function normalizeFallbackChain(value: unknown): string[] {
     result.push(trimmed);
   }
   return result;
-}
-
-function getSeroSettings(settings: Record<string, unknown>): Record<string, unknown> {
-  const raw = settings.sero;
-  if (raw && typeof raw === 'object' && !Array.isArray(raw)) {
-    return { ...(raw as Record<string, unknown>) };
-  }
-  return {};
 }
 
 export function getDefaultModelFallbackChain(): string[] {
